+++ ts/api.ts
@@ -20,51 +20,14 @@ interface IFullfilledXHR extends XMLHttpRequest {
   statusText: string
 }
 
-const APIRateLimit = {
-  // Wait ${coolTime} seconds per toot.
-  coolTime: 20000,
-  isCoolTime: false,
-  // Ratelimit per 1min.
-  rateLimitPublic: 2,
-  rateLimitReply: 30,
-  remainingPublic: 0,
-  remainingReply: 0
-}
-
 export class MastodonAPI {
-  private bearerToken: string
-  private hostName: string
-  private limit = APIRateLimit
-  private _visibility: Visibility = 'direct'
-  // Timer of setInterval() that resets ratelimit.
-  private resetPublicRateLimitID: NodeJS.Timer
-  private resetReplyRateLimitID: NodeJS.Timer
-  // Timer of setInterval() that resets cooltime.
-  private resetCoolTimeID: NodeJS.Timer
+  protected bearerToken: string
+  protected hostName: string
+  protected _visibility: Visibility = 'direct'
 
   constructor (host: string, token: string) {
     this.bearerToken = token
     this.hostName = host
-  }
-
-  public set rateLimitPublic (value: number) {
-    if (!this.limit) return
-    if (value <= 0) return
-    this.limit.rateLimitPublic = value
-  }
-
-  public get rateLimitPublic (): number {
-    return this.limit.rateLimitPublic
-  }
-
-  public set rateLimitReply (value: number) {
-    if (!this.limit) return
-    if (value <= 0) return
-    this.limit.rateLimitReply = value
-  }
-
-  public get rateLimitReply (): number {
-    return this.limit.rateLimitReply
   }
 
   public set visibility (value: Visibility) {
@@ -102,51 +65,7 @@ export class MastodonAPI {
     })
   }
 
-  public setRateLimit (pubvalue?: number, repvalue?: number): void {
-    if (!this.limit) this.limit = APIRateLimit
-
-    if (pubvalue) this.rateLimitPublic = pubvalue
-    this.limit.remainingPublic = this.rateLimitPublic
-
-    if (repvalue) this.rateLimitReply = repvalue
-    this.limit.remainingReply = this.rateLimitReply
-
-    // Reset remaining values to default value per 1min.
-    if (!this.resetPublicRateLimitID) {
-      this.resetPublicRateLimitID = setInterval(() => this.limit.remainingPublic = this.rateLimitPublic, 60000)
-    }
-    if (!this.resetReplyRateLimitID) {
-      this.resetReplyRateLimitID = setInterval(() => this.limit.remainingReply = this.rateLimitReply, 60000)
-    }
-    console.log(`rateLimitPublic: ${this.rateLimitPublic}`)
-    console.log(`rateLimitReply: ${this.rateLimitReply}`)
-  }
-
-  public setCoolTime (value?: number): void {
-    if (!this.limit) this.setRateLimit()
-    if (value) this.coolTime = value
-
-    if (!this.resetCoolTimeID) {
-      this.resetCoolTimeID = setInterval(() => this.limit.isCoolTime = false, this.limit.coolTime)
-    }
-    console.log(`coolTime: ${this.limit.coolTime}`)
-  }
-
-  private set coolTime (value: number) {
-    if (value <= 0) return
-    this.limit.coolTime = value
-  }
-
   public toot (content: string, visibility?: Visibility, replyToID?: string): void {
-    if (replyToID) {
-      const msg = `rateLimitReply: ${this.limit.remainingReply}`
-      if (this.limit && this.limit.remainingReply <= 0) return console.log(msg)
-    } else {
-      const msg = `rateLimitPublic: ${this.limit.remainingPublic}`
-      if (this.limit && this.limit.remainingPublic <= 0) return console.log(msg)
-      if (this.limit.isCoolTime) return console.log(`coolTime: ${this.limit.coolTime}`)
-    }
-
     const data = {
       in_reply_to_id: replyToID ? replyToID : null,
       media_ids: [],
@@ -158,16 +77,6 @@ export class MastodonAPI {
     this.sendToot(data)
     .then((resp) => console.log(resp.status))
     .catch((resp) => console.error(`${resp.status}: ${resp.statusText}`))
-
-    if (!this.limit) return
-    if (replyToID) {
-      this.limit.remainingReply--
-      console.log(`remainingReply: ${this.limit.remainingReply}`)
-    } else {
-      this.limit.remainingPublic--
-      console.log(`remainingPublic: ${this.limit.remainingPublic}`)
-      this.limit.isCoolTime = true
-    }
   }
 
   public favourite (id: string): Promise<IFullfilledXHR> {
@@ -220,7 +129,7 @@ export class MastodonAPI {
     })
   }
 
-  private sendToot (data: ISendToot): Promise<IFullfilledXHR> {
+  protected sendToot (data: ISendToot): Promise<IFullfilledXHR> {
     return new Promise((resolve: (resp: IFullfilledXHR) => void,
                         reject: (err: IFullfilledXHR) => void) => {
       const xhr = new XMLHttpRequest()
